+++ aws-sdk-core/lib/aws-sdk-core.rb
@@ -27,6 +27,7 @@ module Aws
     CloudWatchLogs
     CognitoIdentity
     CognitoSync
+    ConfigService
     DataPipeline
     DirectConnect
     DynamoDB