+++ aws-sdk-core/lib/aws-sdk-core/configservice.rb
@@ -1,0 +1,4 @@
+Aws.add_service(:ConfigService, {
+  api: File.join(Aws::API_DIR, 'ConfigService.api.json'),
+  docs: File.join(Aws::API_DIR, 'ConfigService.docs.json'),
+})